Break up `recap.analyzers.db`
The `db.py` in analyzers was getting pretty big. I am thinking of adding
Pydantic models rather than `dict[str, Any]` for analyzer returns. Before I do
that, I want to make the analyzers more manageable. Functionally, nothing is
changing in this commit; it's just a refactor.
