+++ pyproject.toml
@@ -39,15 +39,15 @@ repository = "https://github.com/recap-cloud/recap"
 recap = "recap.cli:app"
 
 [project.entry-points."recap.analyzers"]
-"db.access" = "recap.analyzers.db:TableAccessAnalyzer"
-"db.column" = "recap.analyzers.db:TableColumnAnalyzer"
-"db.comment" = "recap.analyzers.db:TableCommentAnalyzer"
-"db.foreign_key" = "recap.analyzers.db:TableForeignKeyAnalyzer"
-"db.index" = "recap.analyzers.db:TableIndexAnalyzer"
-"db.location" = "recap.analyzers.db:TableLocationAnalyzer"
-"db.primary_key" = "recap.analyzers.db:TablePrimaryKeyAnalyzer"
-"db.profile" = "recap.analyzers.db:TableProfileAnalyzer"
-"db.view_definitions" = "recap.analyzers.db:TableViewDefinitionAnalyzer"
+"db.access" = "recap.analyzers.db.access:TableAccessAnalyzer"
+"db.column" = "recap.analyzers.db.column:TableColumnAnalyzer"
+"db.comment" = "recap.analyzers.db.comment:TableCommentAnalyzer"
+"db.foreign_key" = "recap.analyzers.db.foreign_key:TableForeignKeyAnalyzer"
+"db.index" = "recap.analyzers.db.index:TableIndexAnalyzer"
+"db.location" = "recap.analyzers.db.location:TableLocationAnalyzer"
+"db.primary_key" = "recap.analyzers.db.primary_key:TablePrimaryKeyAnalyzer"
+"db.profile" = "recap.analyzers.db.profile:TableProfileAnalyzer"
+"db.view_definitions" = "recap.analyzers.db.view_definition:TableViewDefinitionAnalyzer"
 
 [project.entry-points."recap.browsers"]
 db = "recap.browsers.db:DatabaseBrowser"
